Final stages of game. Flocking almost at it's peak. Add final touches.

+++ Assets/Scripts/FollowTarget.cs
@@ -6,14 +6,13 @@ using System.Collections.Generic;
 public class FollowTarget : MonoBehaviour
 {
 
-    public int speed;
+    public float speed;
     public GameObject player;
 
     void Update()
     {
         Vector3 localPosition = player.transform.position - transform.position;
-        localPosition = localPosition.normalized; // The normalized direction in LOCAL space
+        localPosition = localPosition.normalized; // The normalized direction 
         transform.Translate(localPosition.x * Time.deltaTime * speed, localPosition.y * Time.deltaTime * speed, localPosition.z * Time.deltaTime * speed);
-        //transform.LookAt(player);
     }
 }
